Merge in LuaJIT-test-cleanup into the main repo
The tests and benchmarks in the LuaJIT-test-cleanup repo are more or
less complete and with scaffolding added, they can now be called
seamlessly from within the luajit sources using `make check` and `make
bench` respectively.

Since the code of LuaJIT-test-cleanup is in public domain, I have
added the code in here as is with copyright notice changes.  I have
added a CONTRIBUTORS file to list the authors of various commits in
LuaJIT-test-cleanup, bench_lua (from where the benchmark driver was
taken) and LuaJIT itself to allow for a more succint copyright notice
that credits authors in addition to Mike Pall in the COPYRIGHT file.