Patch for PPC64 support
 Create a patch for PPC64 support based on
https://github.com/LuaJIT/LuaJIT/pull/140.
https://bugzilla.redhat.com/show_bug.cgi?id=1591701
 This patch has been rebased to match FPU support

Author: Guy Menanteau <menantea@fr.ibm.com>

Signed-Off-By: Marcin Kościelnicki <koriakin@0x04.net
Signed-Off-By: Siddhesh Poyarekar <siddhesh@gotplt.org>

[ppc] Fix access beyond list in ipairs

The load into TMP2 was incorrectly put into ENDIAN_LE, which made the
subsequent check invalid.

[ppc] Fix typo

[ppc]  Load BASEP4 as much as possible

BASEP4 doesn't seem to get initialized all the time, especially when
BASE is updated because of which programs can crash at random on
ppc32.  Err on the conservative side and set BASEP4 every time BASE_LO
(or BASE_HI for LE) are accessed.

This eventually needs to be tuned optimally.

[ppc] Revert LE code for assert

[ppc] Fix off by one in assert

It ended up reading the first argument twice.

Fix BC_POW on ppc64le