Make jit.prngstate use the new PRNG code
This change splits out the PRNG so that the PRNG used by the JIT
compiler can be overridden without affecting the PRNG for the
allocator.

The API is changed to reflect the wider PRNG data, which is now 256
bits.  jit.prngstate returns an array with 8 32-bit integers and
similarly, it can be overridden with an array of up to 8 32-bit
integers.

Tests and README.md have been updated to reflect this change.