s390x: simplify ceil/floor code
The previous change[1] to this code left quite a bit of dead code
behind, when it switched vm_round from using DIDBR (divide to integer)
to FIDBR (load floating-point integer). Simplify the code by removing
the remains of the previous implementation.

As a side effect, math.ceil() etc. now execute without crashing in QEMU,
which was previously not the case because QEMU doesn't implement DIDBR.

[1]: https://github.com/openresty/luajit2/pull/177